Move avatar tag into wagtailadmin_tags
