+++ wagtail/admin/templates/wagtailadmin/account/change_avatar.html
@@ -1,5 +1,5 @@
 {% extends "wagtailadmin/base.html" %}
-{% load avatar i18n %}
+{% load wagtailadmin_tags i18n %}
 
 {% block titletag %}{% trans "Change profile profile" %}{% endblock %}
 {% block content %}
