+++ wagtail/admin/templates/wagtailadmin/home.html
@@ -1,5 +1,5 @@
 {% extends "wagtailadmin/base.html" %}
-{% load avatar staticfiles i18n %}
+{% load wagtailadmin_tags staticfiles i18n %}
 {% block titletag %}{% trans "Dashboard" %}{% endblock %}
 {% block bodyclass %}homepage{% endblock %}
