+++ wagtail/admin/templates/wagtailadmin/pages/edit.html
@@ -1,6 +1,5 @@
 {% extends "wagtailadmin/base.html" %}
 {% load wagtailadmin_tags %}
-{% load avatar %}
 {% load i18n %}
 {% load l10n %}
 {% block titletag %}{% blocktrans with title=page.get_admin_display_title page_type=content_type.model_class.get_verbose_name %}Editing {{ page_type }}: {{ title }}{% endblocktrans %}{% endblock %}
