+++ wagtail/admin/templates/wagtailadmin/pages/revisions/list.html
@@ -1,4 +1,4 @@
-{% load i18n wagtailadmin_tags avatar %}
+{% load i18n wagtailadmin_tags %}
 {% load l10n %}
 
 <table class="listing">
