+++ wagtail/admin/templates/wagtailadmin/shared/main_nav.html
@@ -1,4 +1,4 @@
-{% load avatar wagtailadmin_tags %}
+{% load wagtailadmin_tags %}
 {% load i18n %}
 <nav class="nav-main">
     <ul>
