+++ wagtail/admin/templates/wagtailadmin/shared/user_avatar.html
@@ -1,4 +1,4 @@
-{% load avatar %}
+{% load wagtailadmin_tags %}
 
 <span class="avatar small">
     <img src="{% avatar_url user size=25 %}" />
