+++ wagtail/contrib/styleguide/templates/wagtailstyleguide/base.html
@@ -1,5 +1,5 @@
 {% extends "wagtailadmin/base.html" %}
-{% load wagtailadmin_tags i18n staticfiles avatar %}
+{% load wagtailadmin_tags i18n staticfiles %}
 
 {% block extra_css %}
     {{ block.super }}
