+++ wagtail/users/templates/wagtailusers/groups/index.html
@@ -1,6 +1,5 @@
 {% extends "wagtailadmin/base.html" %}
 {% load i18n %}
-{% load avatar %}
 {% block titletag %}{% trans "groups" %}{% endblock %}
 {% block extra_js %}
     {{ block.super }}
