+++ wagtail/users/templates/wagtailusers/users/index.html
@@ -1,6 +1,5 @@
 {% extends "wagtailadmin/base.html" %}
 {% load i18n %}
-{% load avatar %}
 {% block titletag %}{% trans "Users" %}{% endblock %}
 {% block extra_js %}
     {{ block.super }}
