+++ wagtail/users/templates/wagtailusers/users/list.html
@@ -1,5 +1,4 @@
-{% load i18n wagtailusers_tags %}
-{% load avatar %}
+{% load i18n wagtailusers_tags wagtailadmin_tags %}
 <table class="listing">
     <thead>
         <tr>
